Support C99 standards (enums, assert)
Fix various errors when -Werror=pedantic flag is enabled:

Cast enum values to int to prevent errors like:
   deps/tinycbor/src/cborconstants_p.h:47:31:\
   error: ISO C restricts enumerator values to range of ‘int’\
   [-Werror=pedantic]
   MajorTypeMask           = ~0U << MajorTypeShift,

"error: ISO C99 does not support" ‘Static_assert’

This effort is needed for RIOT port of iotivity-contrained

Bug-RIOT: https://github.com/RIOT-OS/RIOT/issues/6241

+++ src/cbor.h
@@ -151,8 +151,8 @@ typedef enum CborError {
     CborErrorJsonObjectKeyNotString,
     CborErrorJsonNotImplemented,
 
-    CborErrorOutOfMemory = ~0U / 2 + 1,
-    CborErrorInternalError = ~0U
+    CborErrorOutOfMemory = (int) (~0U / 2 + 1),
+    CborErrorInternalError = (int) ~0U
 } CborError;
 
 CBOR_API const char *cbor_error_string(CborError error);
